+++ libs/ERM_Rendering/Sources/Vulkan/erm/rendering/buffers/IndexBuffer.cpp
@@ -2,7 +2,7 @@
 
 namespace erm {
 
-IndexBuffer::IndexBuffer(Device& device, size_t size, uint32_t count)
+IndexBuffer::IndexBuffer(Device& device, u64 size, u32 count)
 	: DeviceBuffer(device, size, BufferUsage::INDEX_BUFFER
 #ifdef ERM_RAY_TRACING_ENABLED
 					   | BufferUsage::SHADER_DEVICE_ADDRESS | BufferUsage::STORAGE_BUFFER | BufferUsage::ACCELERATION_STRUCTURE_BUILD_INPUT_READ_ONLY
