+++ libs/ERM_Rendering/Sources/Vulkan/erm/rendering/buffers/VertexBuffer.h
@@ -2,12 +2,14 @@
 
 #include "erm/rendering/buffers/DeviceBuffer.h"
 
+#include <erm/math/Types.h>
+
 namespace erm {
 
 class VertexBuffer : public DeviceBuffer
 {
 public:
-	VertexBuffer(Device& device, size_t size);
+	VertexBuffer(Device& device, u64 size);
 };
 
 } // namespace erm
