+++ libs/ERM_Rendering/Sources/common/erm/rendering/buffers/BufferInfo.h
@@ -2,19 +2,21 @@
 
 #include "erm/rendering/Macros.h"
 
+#include <erm/math/Types.h>
+
 namespace erm {
 
 struct BufferInfo
 {
 	BufferInfo(
-		size_t offset = 0,
-		size_t stride = ERM_WHOLE_SIZE)
+		u64 offset = 0,
+		u64 stride = ERM_WHOLE_SIZE)
 		: mOffset(offset)
 		, mStride(stride)
 	{}
 
-	size_t mOffset;
-	size_t mStride;
+	u64 mOffset;
+	u64 mStride;
 };
 
 } // namespace erm
