+++ libs/ERM_Rendering/Sources/common/erm/rendering/enums/PolygonMode.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <erm/math/Types.h>
+
 namespace erm {
 
-enum class PolygonMode : short
+enum class PolygonMode : u8
 {
 	POINT,
 	LINE,
